Update shared prepull normalizer
